Fixing up includes and build for Bspline related file moves
