+++ src/QMCWaveFunctions/BsplineFactory/createRealSingle.cpp
@@ -12,14 +12,14 @@
 #include "qmc_common.h"
 #include <Utilities/ProgressReportEngine.h>
 #include "QMCWaveFunctions/EinsplineSetBuilder.h"
-#include "QMCWaveFunctions/BsplineSet.h"
+#include "QMCWaveFunctions/BsplineFactory/BsplineSet.h"
 #include "QMCWaveFunctions/BsplineFactory/SplineR2RAdoptor.h"
 #include "QMCWaveFunctions/BsplineFactory/HybridRealAdoptor.h"
 #include <fftw3.h>
 #include <QMCWaveFunctions/einspline_helper.hpp>
-#include "QMCWaveFunctions/BsplineReaderBase.h"
-#include "QMCWaveFunctions/SplineAdoptorReaderP.h"
-#include "QMCWaveFunctions/SplineHybridAdoptorReaderP.h"
+#include "QMCWaveFunctions/BsplineFactory/BsplineReaderBase.h"
+#include "QMCWaveFunctions/BsplineFactory/SplineAdoptorReaderP.h"
+#include "QMCWaveFunctions/BsplineFactory/SplineHybridAdoptorReaderP.h"
 
 namespace qmcplusplus
 {
